pyproject.toml
[project]
name = "squimmyai-blog"
version = "0.1.0"
description = "Static site generator for squimmyai.github.io"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "markdown>=3.5",
    "pyyaml>=6.0",
]
